Hide search related resources behind an action button
